chore: split playgrounds test components
