+++ packages/docs/content/0.index.md
@@ -20,7 +20,7 @@ snippet: pnpm add mt-v-safe-html
 ---
 
 #title
-A good alternative for v-html.
+A good guard method before using v-html.
 
 #description
 A lightweight, flexible, and robust XSS sanitizer's Vue directive based on DOMPurify.
